refactor: Registration button on table elements.
