+++ src/admin/pages/ContentType/index.tsx
@@ -1,8 +1,8 @@
 import Table from '@admin/components/elements/Table';
 import Cell from '@admin/components/elements/Table/Cell';
 import { Type } from '@admin/components/elements/Table/Cell/types';
+import { useDocumentInfo } from '@admin/components/utilities/DocumentInfo';
 import buildColumns from '@admin/utilities/buildColumns';
-import { Box } from '@mui/material';
 import { Field } from 'config/types';
 import React from 'react';
 
@@ -13,6 +13,8 @@ const columns = buildColumns(fields, (i: number, row: any, data: any) => (
 ));
 
 const ContentType: React.FC = () => {
+  const documentInfo = useDocumentInfo();
+
   const rows = [
     {
       id: 1,
@@ -28,11 +30,7 @@ const ContentType: React.FC = () => {
     },
   ];
 
-  return (
-    <Box sx={{ width: '100%' }}>
-      <Table columns={columns} rows={rows} />
-    </Box>
-  );
+  return <Table label={documentInfo.label} columns={columns} rows={rows} />;
 };
 
 export default ContentType;
