+++ src/admin/pages/Role/index.tsx
@@ -1,8 +1,8 @@
 import Table from '@admin/components/elements/Table';
 import Cell from '@admin/components/elements/Table/Cell';
 import { Type } from '@admin/components/elements/Table/Cell/types';
+import { useDocumentInfo } from '@admin/components/utilities/DocumentInfo';
 import buildColumns from '@admin/utilities/buildColumns';
-import { Box } from '@mui/material';
 import React from 'react';
 
 const fields = [
@@ -16,6 +16,8 @@ const columns = buildColumns(fields, (i: number, row: any, data: any) => (
 ));
 
 const Role: React.FC = () => {
+  const documentInfo = useDocumentInfo();
+
   const rows = [
     {
       id: 1,
@@ -31,11 +33,7 @@ const Role: React.FC = () => {
     },
   ];
 
-  return (
-    <Box sx={{ width: '100%' }}>
-      <Table columns={columns} rows={rows} />
-    </Box>
-  );
+  return <Table label={documentInfo.label} columns={columns} rows={rows} />;
 };
 
 export default Role;
